requirements.txt
gradio>=4.0.0
google-generativeai
scipy
numpy
